New generate object functions

Closes #504

See merge request emss/aegis!506

+++ src/components/panes/rex/rex-right-rex.tsx
@@ -114,7 +114,9 @@ const RexRightRex: FunctionComponent = () => {
                 <Button
                   icon={faTrashAlt}
                   onClick={() => {
-                    dispatch(thunkDeleteRex({ rexUuid: selectedRex.uuid }));
+                    if (window.confirm("Are you sure you want to delete this Rex?")) {
+                      dispatch(thunkDeleteRex({ rexUuid: selectedRex.uuid }));
+                    }
                   }}
                   toolTip="Delete Rex Item"
                   style={{ width: "30px", fontSize: "0.9em", paddingLeft: "10px" }}
